riscv: remove `vexriscv` crate
The vexriscv crate has been simplified as a result of
https://github.com/xobs/vexriscv-rust/commit/ba269ff348033e68f7f474353993bd5d32d93c3c.
This removed a lot of duplicate code, under the theory that developers
would use `riscv` for common functions, and processor-specific crates
for processor-specific functions.

This patchset switches back to using `riscv`.

This is required as this no longer builds on nightly due to the renaming of `asm!()` to `llvm_asm!()`,
which necessitates reworking the `vexriscv` crate.

Signed-off-by: Sean Cross <sean@xobs.io>